Rename ProcessGroup_ez.h to ProcessGroup_other.h
